Moved the TractionField and PressureField class definitions to separate header file for clarity, while renaming Function.C to TractionField.C. Avoid including Function.h in IntegrandBase.h - use forward declarations.
